refactor create shop utilities
